Update flutter_clean_calendar.dart
This change solve the a problem a found on the expandable date. No matter if you change the locale to other language, this date was always shown in English.
I changed it, using DateFormat instead.

Besides, I add a new property called expandableDateFormat that allow the user choose a format. The default format is "EEEE MMMM dd, yyyy"