#3 Added a dedicated class for clean calendar events
This class allows it to set certain options to an event. Only the title (summary) is mandatory, all other parameters are optional. This is a breaking change.

#3 Example adapted for usage of new events class
#3 Example edited to make use of the CleanCalendarEvent class
The example now imports the CleanCalendarEvent and uses the parameters of that class to show a more detailed event list for the selected day
#3 The dots representing an event now take the color of the color property of the CleanCanlenderEvent
If the CleanCalendarEvent has no color property set, you can set the color with the eventColor property of the CalendarTile
#3 Example of how to use the color property of the CleanCalenderEvent
#3 New property to set the string of the Today button on top of the screen
By setting this property it is possible to set the caption of the Today button in local langauges
#3 New property for callback function for changing to expanded view
Eliminated double quotes and replaced them with single quotes
#3 Example now forces the list of today's events being shown on load
This is done by calling _handleNewDate in initState method